requirements.txt
python-constraint
jupyterlab
matplotlib
numpy
pytest